Refactor: Add CRUD operations for Author entity in AuthorBL.java
This commit adds the CRUD (Create, Read, Update, Delete) operations for the Author entity in the AuthorBL.java file. It includes methods to get an author by ID, add a new author, edit an existing author, and delete an author. The code has been refactored to improve the readability and maintainability of the codebase.